split the test for db and facial, changed load known faces to pandas dataframe and returning ID with nparray for encoding values. test the interface and see what you get back when you confirm attendance.
